% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoastlineFractalDimension.R
\name{BoxesFD}
\alias{BoxesFD}
\title{BoxesFD}
\usage{
BoxesFD(BinputPath, netPath, outputPath, year, r, pearsonValue, writeF, showF)
}
\arguments{
\item{BinputPath}{All origin shoreline files path}

\item{netPath}{All fishnet files path}

\item{outputPath}{All results will be exported here}

\item{year}{R vector object, which represent your study time}

\item{r}{R vector object, which represent your study scale}

\item{pearsonValue}{The Pearson coefficient of your input data}

\item{writeF}{Exporting Function's result}

\item{showF}{Drawing Function's result}
}
\value{
An .xlsx file containing the results of the coastline fractal dimension
}
\description{
Calculating the Coastline Fractal Dimension with Boxes methods
}
\examples{

BinputPath = list.files(system.file('extdata', package = 'CoastlineFD'),full.names = TRUE)[1]
netPath = list.files(system.file('extdata', package = 'CoastlineFD'),full.names = TRUE)[3]
outputPath = paste0(system.file('extdata', package = 'CoastlineFD'), "/FD1985_1986.xlsx")

BoxesFD(
  BinputPath,
  netPath,
  outputPath,
  c(1985:1986),
  c(300, 600, 900, 1000, 1050, 1100),
  0.00,
  FALSE,
  TRUE
)

}
