% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdc.R
\name{FDC}
\alias{FDC}
\title{FDC}
\usage{
FDC(inputPath,netPath,outputPath,year,...)
}
\arguments{
\item{inputPath}{All density shoreline files path, up to the last folder}

\item{netPath}{All fishnet files path, up to the last folder}

\item{outputPath}{FD results will be exported here, up to the file name, only support .xlsx export}

\item{year}{R vector object, which represent your study time}

\item{r}{R vector object, which represent your study scale}

\item{pearsonValue}{The Pearson coefficient of your input data}

\item{flag}{Drawing Dividers_Function's result}
}
\description{
Calculating Fractal Dimension of coastline
}
\details{
You must densify line shapefiles by QGIS or something else at first
}
