% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoastlineFractalDimension.R
\name{FD}
\alias{FD}
\title{CoastlineFD}
\usage{
FD(DinputPath, BinputPath, netPath, outputPath, year, r, pearsonValue, showFlag)
}
\arguments{
\item{DinputPath}{All density shoreline files path, up to the last folder}

\item{BinputPath}{All shoreline files path, up to the last folder}

\item{netPath}{All fishnet files path, up to the last folder}

\item{outputPath}{FD results will be exported here, up to the file name}

\item{year}{R vector object, which represent your study time}

\item{r}{R vector object, which represent your study scale}

\item{pearsonValue}{The Pearson coefficient of your input data}

\item{showFlag}{Drawing Function's result}
}
\description{
Calculate the Coastline Fractal Dimension
}
